follow convention to write to input if DESTROY_INPUT is set.
This fixed all cases but the PADDED r2C and likely the shifted transforms.

The problem with the padded r2c transforms is that somehow the last axis local_ni
is not padded. It should be a trivial fix?

I am lost here, because the size is actually computed outside the remap routines
by partrafo.c The routines looked generic enough to support any decompositions,
including 2d on 2d and 3d on 3d.