roughly follow the heuristics in remap_3dto2d.c
It appears that we shall prefer out of place global_transp
as much as possible.

The FFTW_PRESERVE_INPUT flag is added to signify that the transform
must be preserving input; it should have been added automatically.

Also expanding the code branches to make it easier to parse
what is actually done in different cases.